Move ClientReconnectIT test to separate execution. Tune pom.xml
