+++ src/test/java/com/github/kpavlov/jreactive8583/it/ClientReconnectIT.java
@@ -1,11 +1,13 @@
-package com.github.kpavlov.jreactive8583.example;
+package com.github.kpavlov.jreactive8583.it;
 
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.TimeUnit;
 
 import static org.awaitility.Awaitility.await;
 
+@Tag("reconnect")
 public class ClientReconnectIT extends AbstractIT {
 
     @Test
@@ -19,7 +21,7 @@ public class ClientReconnectIT extends AbstractIT {
         await().alias("client was disconnected").until(() -> (!client.isConnected()));
 
         //then
-        TimeUnit.SECONDS.sleep(7);
+        TimeUnit.SECONDS.sleep(3);
         server.init();
         server.start();
         await().alias("server started").until(server::isStarted);
