+++ src/test/resources/application-test.properties
@@ -1,4 +1,4 @@
 # suppress inspection "UnusedProperty" for whole file
 iso8583.connection.host=127.0.0.1
 iso8583.connection.port=9876
-iso8583.connection.idleTimeout=5
+iso8583.connection.idleTimeout=2
